chore: split transformers to have an mdxish tests subdir
