+++ __tests__/transformers/mdxish/preprocess-redos-attack.test.ts
@@ -1,4 +1,4 @@
-import { preprocessJSXExpressions } from '../../processor/transform/mdxish/preprocess-jsx-expressions';
+import { preprocessJSXExpressions } from '../../../processor/transform/mdxish/preprocess-jsx-expressions';
 
 describe('ReDoS Attack Vectors', () => {
   it('should handle basic attack pattern without hanging', () => {
